requirements.txt
dj-database-url==0.5.0
Django==1.11.16
django-cors-headers==2.4.0
django-heroku==0.3.1
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2==2.7.5
pytz==2018.5
whitenoise==4.1
